ASoC: SOF: sof-client-probes: Add available points_info(), IPC4 only
Add another debugfs file, "probe_points_available", that shows all the
available probe points in the SOF FW at the time of query. The probe
points are there only when an active SOF stream exists in the
system. However, the stream identifiers are persistent in the sense
that the same probe point identifiers always appear with the same
playback or capture command in the same system configuration.

The output, when reading "probe_points_available", may look like this:

16777220,0,256 host-copier.0.playback output 1 buf idx 0 (probed)
6,0,256	gain.1.1 input buf idx 0 (probed)
16777222,0,0	gain.1.1 output buf idx 0
2,0,0	mixin.1.1 input buf idx 0
16777218,0,0	mixin.1.1 output buf idx 0
3,0,0	mixout.2.1 input buf idx 0
16777219,0,0	mixout.2.1 output buf idx 0
65542,0,0	gain.2.1 input 0 buf idx 0
16842758,0,0	gain.2.1 output buf idx 0
15,0,0	smart_amp.2.1 input buf idx 0
16777231,0,0	smart_amp.2.1 output buf idx 0
65540,0,0	dai-copier.SSP.NoCodec-0.playback input buf idx 0

The triplet at the beginning of a line can be copy-pasted as such to
"probe_points" debugfs file for adding a probe point. The rest of the
line tries to give human readable explanation of what this probe point
is.

Signed-off-by: Jyri Sarha <jyri.sarha@linux.intel.com>